Move const match iterator into seperate file
