+++ src/core/stores/tatami/meson.build
@@ -3,4 +3,5 @@ core_sources += ['src/core/stores/tatami/sequential_block_group.cpp']
 core_sources += ['src/core/stores/tatami/tatami_list.cpp']
 core_sources += ['src/core/stores/tatami/location.cpp']
 core_sources += ['src/core/stores/tatami/tatami_store.cpp']
+core_sources += ['src/core/stores/tatami/const_match_iterator.cpp']
