+++ src/core/stores/tatami/sequential_block_group.hpp
@@ -5,6 +5,7 @@
 #include "core/core.hpp"
 #include "core/id.hpp"
 #include "core/serialize.hpp"
+#include "core/stores/tatami/const_match_iterator.hpp"
 
 enum class MatchType;
 class TournamentStore;
@@ -12,27 +13,6 @@ struct BlockLocation;
 
 class SequentialBlockGroup {
 public:
-    // Iterates non-hidden matches in a sequential block group
-    class ConstMatchIterator {
-    public:
-        ConstMatchIterator & operator++();
-        CombinedId operator*();
-        bool operator!=(const ConstMatchIterator &other) const;
-        bool operator==(const ConstMatchIterator &other) const;
-
-    private:
-        ConstMatchIterator(const TournamentStore &tournament, const SequentialBlockGroup &group, size_t currentBlock, size_t currentMatch);
-        void loadMatch();
-
-        const TournamentStore &mTournament;
-        const SequentialBlockGroup &mGroup;
-        const CategoryStore *mCurrentCategory;
-        MatchType mCurrentType;
-        size_t mCurrentBlock;
-        size_t mCurrentMatch;
-        friend class SequentialBlockGroup;
-    };
-
     SequentialBlockGroup();
 
     void eraseBlock(std::pair<CategoryId, MatchType> block);
